+++ Source/Mosa.Kernel.x86/Screen.cs
@@ -182,10 +182,10 @@ namespace Mosa.Kernel.x86
 		/// <param name="val">The val.</param>
 		/// <param name="digits">The digits.</param>
 		/// <param name="size">The size.</param>
-		public static void Write(ulong val, byte digits, int size)
+		public static void Write(uint val, byte digits, int size)
 		{
 			uint count = 0;
-			ulong temp = val;
+			uint temp = val;
 
 			do
 			{
